LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/size-feature-evaluation-expected.txt

Test

FAIL size(width < 100px) (.horizontal) assert_equals: expected "" but got "true"
PASS size(width >= 100px) (.horizontal)
PASS size(min-width: 100px) (.horizontal)
FAIL size(min-width: 101px) (.horizontal) assert_equals: expected "" but got "true"
PASS size(max-width: 100px) (.horizontal)
FAIL size(max-width: 99px) (.horizontal) assert_equals: expected "" but got "true"
FAIL size(height < 200px) (.horizontal) assert_equals: expected "" but got "true"
PASS size(height >= 200px) (.horizontal)
PASS size(min-height: 200px) (.horizontal)
FAIL size(min-height: 201px) (.horizontal) assert_equals: expected "" but got "true"
PASS size(max-height: 200px) (.horizontal)
FAIL size(max-height: 199px) (.horizontal) assert_equals: expected "" but got "true"
FAIL size(inline-size < 100px) (.horizontal) assert_equals: expected "" but got "true"
PASS size(inline-size >= 100px) (.horizontal)
PASS size(min-inline-size: 100px) (.horizontal)
FAIL size(min-inline-size: 101px) (.horizontal) assert_equals: expected "" but got "true"
PASS size(max-inline-size: 100px) (.horizontal)
FAIL size(max-inline-size: 99px) (.horizontal) assert_equals: expected "" but got "true"
FAIL size(block-size < 200px) (.horizontal) assert_equals: expected "" but got "true"
PASS size(block-size >= 200px) (.horizontal)
PASS size(min-block-size: 200px) (.horizontal)
FAIL size(min-block-size: 201px) (.horizontal) assert_equals: expected "" but got "true"
PASS size(max-block-size: 200px) (.horizontal)
FAIL size(max-block-size: 199px) (.horizontal) assert_equals: expected "" but got "true"
FAIL size(orientation: landscape) (.horizontal) assert_equals: expected "" but got "true"
PASS size(orientation: portrait) (.horizontal)
PASS size(aspect-ratio: 1/2) (.horizontal)
FAIL size(aspect-ratio: 2/1) (.horizontal) assert_equals: expected "" but got "true"
FAIL size(width < 100px) (.vertical) assert_equals: expected "" but got "true"
PASS size(width >= 100px) (.vertical)
PASS size(min-width: 100px) (.vertical)
FAIL size(min-width: 101px) (.vertical) assert_equals: expected "" but got "true"
PASS size(max-width: 100px) (.vertical)
FAIL size(max-width: 99px) (.vertical) assert_equals: expected "" but got "true"
FAIL size(height < 200px) (.vertical) assert_equals: expected "" but got "true"
PASS size(height >= 200px) (.vertical)
PASS size(min-height: 200px) (.vertical)
FAIL size(min-height: 201px) (.vertical) assert_equals: expected "" but got "true"
PASS size(max-height: 200px) (.vertical)
FAIL size(max-height: 199px) (.vertical) assert_equals: expected "" but got "true"
FAIL size(block-size < 100px) (.vertical) assert_equals: expected "" but got "true"
PASS size(block-size >= 100px) (.vertical)
PASS size(min-block-size: 100px) (.vertical)
FAIL size(min-block-size: 101px) (.vertical) assert_equals: expected "" but got "true"
PASS size(max-block-size: 100px) (.vertical)
FAIL size(max-block-size: 99px) (.vertical) assert_equals: expected "" but got "true"
FAIL size(inline-size < 200px) (.vertical) assert_equals: expected "" but got "true"
PASS size(inline-size >= 200px) (.vertical)
PASS size(min-inline-size: 200px) (.vertical)
FAIL size(min-inline-size: 201px) (.vertical) assert_equals: expected "" but got "true"
PASS size(max-inline-size: 200px) (.vertical)
FAIL size(max-inline-size: 199px) (.vertical) assert_equals: expected "" but got "true"
FAIL size(orientation: landscape) (.vertical) assert_equals: expected "" but got "true"
PASS size(orientation: portrait) (.vertical)
PASS size(aspect-ratio: 1/2) (.vertical)
FAIL size(aspect-ratio: 2/1) (.vertical) assert_equals: expected "" but got "true"
